+++ test/service/migrations/mysql/verify/user_details.sql
@@ -1,0 +1,7 @@
+-- Verify restfulness:user_details on mysql
+
+BEGIN;
+
+    select id,country,user_id from user_details where false;
+
+ROLLBACK;