+++ test/service/migrations/pg/revert/user_details.sql
@@ -1,0 +1,7 @@
+-- Revert restfulness:user_details from pg
+
+BEGIN;
+
+    drop table user_details;
+
+COMMIT;