+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## FUTURE
+
+* [ADDED] - `resource.hasOne()` method to define 1x1 relationship
+
 ## 0.6.1
 
 * [ADDED] - `route.reducesDatasetBy(Array<string>)` method